+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Add support for libraries.
 - Add support for abstract contracts and interfaces.
+- Generate a constructor that invokes all uninitialized parents.
 
 ## 0.1.0